Revamped config
