+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ontoform"
-version = "25.0.0-rc.3"
+version = "25.0.0-rc.4"
 description = "Open Targets pipeline converter tool"
 readme = "README.md"
 requires-python = ">=3.12"
@@ -20,6 +20,10 @@ ontoform = "ontoform.main:main"
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.coverage.report]
+omit = ["tests/**"]
+exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests/pis"]
