+++ src/ontoform/steps/expression.py
@@ -6,13 +6,13 @@ expression = Step(
     name='expression',
     transformations=[
         FileTransformation(
-            src_path='input/expression-inputs/normal_tissue.tsv.zip',
-            dst_path='input/expression-inputs/normal_tissue.tsv.gz',
+            src_path='input/expression/normal_tissue.tsv.zip',
+            dst_path='intermediate/expression/normal_tissue.tsv.gz',
             transformer=NormalTissueTransformer,
         ),
         FileTransformation(
-            src_path='input/expression-inputs/map_with_efos.json',
-            dst_path=lambda _, f: f'input/expression-inputs/tissue-translation-map.{stem(f)}',
+            src_path='input/expression/map_with_efos.json',
+            dst_path=lambda _, f: f'intermediate/expression/tissue-translation-map.{stem(f)}',
             transformer=TissueTransformer,
         ),
     ],
