+++ src/ontoform/steps/openfda.py
@@ -8,8 +8,8 @@ openfda = Step(
     name='openfda',
     transformations=[
         GlobTransformation(
-            src_prefix='input/fda-inputs',
-            dst_path=lambda _, f: f'input/fda-inputs/{uuid4()}.{stem(f)}',
+            src_prefix='input/openfda',
+            dst_path=lambda _, f: f'intermediate/openfda/{uuid4()}.{stem(f)}',
             glob='**/*.zip',
             transformer=OpenFdaTransformer,
         ),
